微调macOS下的表现

+++ src/hostmgr.py
@@ -27,7 +27,7 @@ FN_HOSTS = r'C:\Windows\System32\drivers\etc\hosts'
 class hostmgr:
     def __init__(self) -> None:
         self.logger=setup_logger()
-        if (os.path.isfile(FN_HOSTS) == False):
+        if (os.path.isfile(FN_HOSTS) == False and sys.platform.startswith('win')):
             self.logger.warning(f"Hosts文件不存在，尝试创建中...")
             try:
                 open(FN_HOSTS, 'w').close()
